requirements.txt
# requirements.txt
streamlit
pandas
numpy
scikit-learn
lightgbm
torch
mecab-python3
joblib
datasets
